app/requirements.txt
python-dotenv
Django==3.1.1
djangorestframework
psycopg2-binary==2.8.6
gunicorn==20.0.4
